+++ packages/docs/.storybook/preview.tsx
@@ -29,6 +29,7 @@ const preview: Preview = {
 			);
 		}
 	],
+
 	parameters: {
 		controls: {
 			matchers: {