+++ packages/react/src/index.tsx
@@ -1,12 +1,13 @@
 "use client";
 
+export * from "./components/avatar";
+export * from "./components/box";
 export * from "./components/button";
+export * from "./components/checkbox";
 export * from "./components/heading";
+export * from "./components/multi-step";
 export * from "./components/text";
-// export * from "./components/Avatar";
-// export * from "./components/Checkbox";
-// export * from "./components/MultiStep";
-// export * from "./components/TextArea";
-// export * from "./components/TextInput";
+export * from "./components/text-area";
+export * from "./components/text-input";
 
 export * from "./styles";